+++ app/src/views/components/poweroff/style.scss
@@ -1,0 +1,3 @@
+.background-fill {
+  opacity: 0;
+}
