+++ app/src/views/pages/game/index.js
