+++ app/src/views/pages/game/style.scss
