+++ app/src/views/pages/loading/style.scss
